Changes how masks are generated and stored
All positions of a mask are now generated and stored with it, so that
when matching a mask to a board, we can iterate through the vector of
masks to count total matches.

Furthermore, there is a vector for any given length of matches, so that
we can check individual lengths at a time.