Refactors board structure
Now, instead of keeping an X board and a O board, we keep a board for
the pieces of the player who just moved and a board for all of the
pieces at once. We switch between players with the XOR operation.

This fixes the masks matching ignoring pieces of the same side they are
checking with the anti mask.